Make autocommit recognize when unstaged files are present
I think this is why I had that diff-index thing, I'd misread
https://stackoverflow.com/a/5139672/34799 to think diff-index
covered all changes, not just changes in tracked files.
Anyway, looks like Git really doesn't have a test for this,
so I'll use this, which is like a plumbing-er version of testing
whether `git status --porcelain` is empty.