Reduce shotclock pressure to 25 seconds
Per the thoughts documented in being-messy.md: I was initially going to reduce this to an outright *60* seconds, but decided to be a little more conservative in opening this aperture, feeling that 25 seconds is usually enough to decide if I've got more to say before it's published, and that 60 might be overkill (leaving me too much time to chicken out altogether). I'm still open to the idea of expanding this window further: I just want to try the experiment a little less drastically first, to see what the effects and side-effects will be.